git: Add move-to
Moves the current commits not present in upstream to a new branch.
This is helpful when you commit something in master that should really
be in a different branch.

For example, you commit a few commits to master, but you want them on
_feature-branch_, you can just use:

  git move-to feature-branch

This will create that branch with your commits, and reset the previous
branch to its upstream.